Fix Test Running and Environment
Make it possible to run `bin/rake` from a fresh clone of the project and have the specs run. Removes unused dependencies for testing and adds required dependencies for testing.

We are requiring the dependencies in the Gemfile for now so that we do not affect customer's ability to install the `spree_retailops` gem.